api: Lamp values are floats now.

+++ VisualPinball.Unity/VisualPinball.Unity/Game/Engine/DefaultGamelogicEngine.cs
@@ -294,7 +294,7 @@ namespace VisualPinball.Unity
 			OnCoilChanged?.Invoke(this, new CoilEventArgs(n, value));
 		}
 
-		public void SetLamp(string id, int value, bool isCoil = false, LampSource source = LampSource.Lamp)
+		public void SetLamp(string id, float value, bool isCoil = false, LampSource source = LampSource.Lamp)
 		{
 			OnLampChanged?.Invoke(this, new LampEventArgs(id, value, isCoil, source));
 		}
